feat(pathfinding): add pathfinding visualizer page
This commit adds a new file `page.tsx` to the `src/app/projects/pathfinding` directory. The page component renders a pathfinding visualizer with a title, breadcrumbs, date, badges, description, and an image. The purpose of this page is to showcase the implementation of four different graph-searching algorithms: A star, BFS, DFS, and Dijkstra. The algorithms are tested on randomly generated maps of different sizes to compare their performance and verify their theoretical big-O. The implementation of the algorithms can be found [here](http://webkutil.cz/algorithms/) and the research paper can be read [here](https://rucforsk.ruc.dk/ws/files/68132039/BP2_Group_1_Pathfinding__new___4_.pdf).

feat(zelda): update description of zelda project

This commit updates the description of the Zelda project in the `page.tsx` file